pyproject.toml
[tool.poetry]
name = "atena"
version = "0.1.0"
description = ""
authors = ["Phawit Pornwattanakul <me@helloyeew.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
python-decouple = "^3.8"
django-admin-interface = "^0.26.0"
django-crispy-forms = "^2.0"
crispy-bootstrap5 = "^0.7"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.6"
django-cryptography = "^1.1"
djangorestframework = "^3.14.0"
boto3 = "^1.28.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
